scripts: mkimage/uboot_helper: add 'extra_dtbs'
There are a some boards which have different revisions and therefore require
a special device tree to make all hardware work for them. In order to not have to
build an extra image for all of them this adds a new 'extra_dtbs' property to
uboot_helper where these device trees can be specified in so that they can get
copied to /flash partition as well.